backend/requirements.txt
Django==2.2.19
pkg_resources==0.0.0
pytz==2022.1
sqlparse==0.4.2
djangorestframework==3.12.4
django-filter
djoser
